% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as01.R
\name{as01}
\alias{as01}
\title{Convert pixel values to 0 and 1}
\usage{
as01(Stack)
}
\arguments{
\item{Stack}{an image stack object}
}
\value{
An array of the same dimensions as \code{Stack} with pixel values changed to 0 and 1.
}
\description{
Convert the pixel values of a binarised image stack to zeros and ones.
}
\author{
Alessio Veneziano
}
