% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleSkeleton.R
\docType{data}
\name{exampleSkeleton}
\alias{exampleSkeleton}
\title{example 3D topological skeleton}
\usage{
data(exampleSkeleton)
}
\description{
3D topological skeleton of the cancellous bone of the mandibular condyle of Hylobates lar (specimen MCZ:Mamm:41421 in morphosource.org). The cancellous bone was extracted using the function \code{splitBone}. The skeleton was calculated using the Amira software (FEI visualization).
}
\author{
Alessio Veneziano
}
\keyword{indianaBones}
