% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleStack.R
\docType{data}
\name{exampleStack}
\alias{exampleStack}
\title{example 3D image stack}
\usage{
data(exampleStack)
}
\description{
Binarized 3D image stack of the mandibular condyle of Hylobates lar (specimen MCZ:Mamm:41421 in morphosource.org)
}
\author{
Alessio Veneziano
}
\keyword{indianaBones}
