% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invertStack.R
\name{invertStack}
\alias{invertStack}
\title{Invert image stack}
\usage{
invertStack(Stack)
}
\arguments{
\item{Stack}{an image stack object}
}
\value{
An array of the same dimensions as \code{Stack} with inverted pixel values.
}
\description{
Inverts all the pixel values in an image stack, creating a negative of the images.
}
\examples{
#Invert pixel values of binary volume and plot 2D image
data(exampleStack)
invStack<-invertStack(exampleStack)
image(exampleStack[,,20])
image(invStack[,,20])

}
\author{
Alessio Veneziano
}
