% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is01.R
\name{is01}
\alias{is01}
\title{Check 01 pixels in image stack}
\usage{
is01(Stack)
}
\arguments{
\item{Stack}{an image stack object}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{Stack} has 01 pixels.
}
\description{
Check if the pixel values of a binarised image stack are zeros and ones.
}
\author{
Alessio Veneziano
}
