% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSphere.R
\name{makeSphere}
\alias{makeSphere}
\title{Spheric Volume}
\usage{
makeSphere(nvoxel, offset = 2)
}
\arguments{
\item{nvoxel}{number of voxels along the diameter of the sphere.}

\item{offset}{number of voxels of the border around the cubic volume}
}
\value{
\item{coords}{a matrix of voxel coordinates of the sphere.}
\item{volume}{a binarised 3D array of the volume of sphere.}
}
\description{
Generates a Spheric volume of user's specified size.
}
\details{
\code{makeSphere} is used by the function \code{makeVOI} to extract VOIs from a 3D volume.
}
\examples{
#Create a volume of a 21 voxels diameter sphere with 2 voxels of offset
sph<-makeSphere(21,2)

#Plot the 3D coords of the sphere
require(rgl)
points3d(sph$coords,col="grey",size=2)

#Plot a slice of the volume (the sphere is in white pixels)
image(sph$volume[,,10],col=grey(0:32/32))

}
\author{
Alessio Veneziano
}
