% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAmiraSkeleton.R
\name{readAmiraSkeleton}
\alias{readAmiraSkeleton}
\title{Read topological skeleton}
\usage{
readAmiraSkeleton(path, removeLoops = T, removeConn = F, maxConn = 10)
}
\arguments{
\item{path}{path of the topological skeleton file (exported from Amira/Avizo in ASCII format)}

\item{removeLoops}{logical: if TRUE, loop skeleton branches (same initial and ending node) are removed}

\item{removeConn}{logical: if TRUE, removes the nodes (and linked branches) connected to more than \code{maxConn} branches}

\item{maxConn}{maximum number of branches connected to each node; ignored if \code{removeConn = FALSE}}
}
\value{
A list including: node coordinates (VertexCoordinates), branch coordinates (EdgePointsCoordinates), node-to-node connectivity (EdgeConnectivity),
number of 3D points constituting each branch (NumEdgePoints) and thickness at each branch point (Thickness).
}
\description{
Read topological skeleton file exported from the Amira/Avizo software in ASCII format.
}
\author{
Alessio Veneziano
}
