% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTiffStack.R
\name{readTiffStack}
\alias{readTiffStack}
\title{Read TIFF image stack}
\usage{
readTiffStack(path, cores = 2)
}
\arguments{
\item{path}{path of the folder where the TIFF images are stored}

\item{cores}{number of cores for parallel computation (default=2)}
}
\value{
An array with dimensions depending on the number of channels in the images. The first two dimensions indicate the image width and height in pixels, and the last dimension is the number of images in the stack.
}
\description{
Reads TIFF image stacks into an array. It uses the function \code{readTIFF} in the package 'tiff' and runs in parallel using the packages 'foreach' and 'doParallel'.
}
\author{
Alessio Veneziano
}
