% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleStack.R
\name{scaleStack}
\alias{scaleStack}
\title{Scale image stack}
\usage{
scaleStack(Stack, newMin = 0, newMax = 1)
}
\arguments{
\item{Stack}{an image stack object}

\item{newMin}{the new minimum pixel value (default=0)}

\item{newMax}{the new maximum pixel value (default=1)}
}
\value{
An array of the same dimensions as \code{Stack} with scaled pixel values.
}
\description{
Scales the pixel values of an image stack between a new maximum and minimum values.
}
\author{
Alessio Veneziano
}
