% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelConnectivity.R
\name{skelConnectivity}
\alias{skelConnectivity}
\title{Skeleton connectivity}
\usage{
skelConnectivity(skel)
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}
}
\value{
A vector of the number of branches connected to each node. The vector has length equal to the number of nodes in the topological skeleton.
}
\description{
Calculates the connectivity of the topological skeleton as the number of branches connected to each node.
}
\examples{
#Calculate skeleton connectivity
data(exampleSkeleton)
conn<-skelConnectivity(exampleSkeleton)

}
\references{
Veneziano A, Cazenave M, Alfieri F, Panetta D, Marchi D. 2021. Novel strategies for the characterization of cancellous bone morphology: Virtual isolation and analysis. American Journal of Physical Anthropology.
}
\author{
Alessio Veneziano
}
