% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelDensity.R
\name{skelDensity}
\alias{skelDensity}
\title{Skeleton density}
\usage{
skelDensity(skel, method = c("node", "edge"), Scale = T, nxyz = c(50, 50, 50))
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}

\item{method}{defines if the density is calculated based on the skeleton nodes ('node') or branches ('edge')}

\item{Scale}{logical: if TRUE, the noes and branch coordinates are scaled on their overall variance}

\item{nxyz}{a vector of length 3 indicating the dimensions of the 3D grid for the estimation of the kernel density (default = 50x50x50)}
}
\value{
\item{Density}{a vector of densities estimated at each node (if \code{method = 'node'}; unit: node per cube cm) or branch point (if \code{method = 'edge'}; unit: branch point per cube cm).}
\item{map}{an array of dimension \code{nxyz} with the estimated kernel density.}
}
\description{
Calculates 3D density of nodes and branches of the topological skeleton based on a Kernel density estimate,
using the \code{kde3d} function in the package 'misc3d'. Density is calculated as point per cube centimeter.
}
\examples{
#Calculate and visualize skeleton node density
data(exampleSkeleton)
dens<-skelDensity(exampleSkeleton,method="node")
  image(dens$map[,,25])

}
\references{
Venables, WN and Ripley, BD. 2002. Modern Applied Statistics with S. Fourth edition. Springer.
}
\author{
Alessio Veneziano
}
