% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelDirection.R
\name{skelDirection}
\alias{skelDirection}
\title{Skeleton angle and direction}
\usage{
skelDirection(skel, ax = 1)
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}

\item{Axis}{the axis to use as reference for the calculation of the branch angle.}
}
\value{
\item{angles}{a vector of branch angles, of length equal to the number of branches.}
\item{directions}{a matrix of branch vector directions.}
\item{refAxis}{a matrix of the coordinates of the reference axis.}
\item{resDir}{a vector of the main branch direction resulting from vector sum.}
\item{resAng}{the main branch angle calculated in radians.}
}
\description{
Calculates the direction of the branches of the topological skeleton. The vector sum of all directions is used to calculate
an overall angle of the skeleton branches with respect to a reference axis.
}
\details{
The 3D angle is calculated in radians between a reference axis and the unitary resultant of all branch directions
obtained by vector sum in 3D. \code{Axis} can be provided by the user as a matrix with starting and ending coordinates of
the axis. Alternatively, if \code{Axis} is set to 1, 2 or 3, the reference axis is calculated as the principal component
axes of the skeleton node coordinates using \code{getAxis}.
}
\references{
Veneziano A, Cazenave M, Alfieri F, Panetta D, Marchi D. 2021. Novel strategies for the characterization of cancellous bone morphology: Virtual isolation and analysis. American Journal of Physical Anthropology.
}
\author{
Alessio Veneziano
}
