% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelThickness.R
\name{skelThickness}
\alias{skelThickness}
\title{Skeleton thickness}
\usage{
skelThickness(skel)
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}
}
\value{
A vector of branch thickness calculated at each coordinate making up the branch.
}
\description{
Extracts branch thickness from the topological skeleton. The thickness is recorded at each point of the branch.
}
\examples{
#Extract thickness information from the topological skeleton
data(exampleSkeleton)
dens<-skelThickness(exampleSkeleton)

}
\author{
Alessio Veneziano
}
