% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelTortuosity.R
\name{skelTortuosity}
\alias{skelTortuosity}
\title{Skeleton tortuosity}
\usage{
skelTortuosity(skel, cores = 2)
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}

\item{cores}{number of cores for parallel computation (default=2)}
}
\value{
A vector of branch tortuosity, of length equal to the number of branches in the topological skeleton.
}
\description{
Calculates branch tortuosity of the topological skeleton.
}
\details{
Tortuosity of a segment is the ratio between its length and the linear distance between its starting and ending points. In the topological
skeleton, tortuosity is calculated as the ratio between the the branch length and the linear distance between the nodes it connects. Tortuosity
ranges from 1 to infinity.
}
\examples{
#Calculate branch tortuosity of the topological skeleton
data(exampleSkeleton)
tort<-skelTortuosity(exampleSkeleton)

}
\references{
Roque WL, & Alberich-Bayarri A. 2015. Tortuosity influence on the trabecular bone elasticity and mechanical competence. In Jorge RN, Tavares JM (Eds) Developments in Medical Image Processing and Computational Vision, pp 173-191. Springer, Cham
}
\author{
Alessio Veneziano
}
