% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorAngle.R
\name{vectorAngle}
\alias{vectorAngle}
\title{Calculate angle between two vectors}
\usage{
vectorAngle(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
Angle between vectors in radians.
}
\description{
Calculates angle between vectors.
}
\author{
Alessio Veneziano
}
