% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOI.R
\name{VOI}
\alias{VOI}
\title{Extract VOI from 3D image Stack}
\usage{
VOI(Stack, shape = c("sphere", "cube"), size, pos)
}
\arguments{
\item{Stack}{an image stack object}

\item{shape}{maximum number of 3D pixels (voxels) considered as isolated islands}

\item{size}{for a spheric VOI, the number of voxels along the diameter of the sphere; for a cubic VOI, number of voxels along the side of the cube.}

\item{pos}{center position of the VOI.}
}
\value{
An array of dimensions size x size x size containing the VOI subset from the image stack.
}
\description{
This function extracts a Volume Of Interest (VOI) of the desired size and shape from a 3D image stack.
}
\details{
\code{VOI} is a simple function to extract a VOI from a 3D image stack. It needs careful positioning.
}
\examples{
#Extract a spherical VOI at the barycenter of a Stack
data(exampleStack)
pos<-apply(which(Stack!=0,T),2,mean)
voi<-VOI(exampleStack,"sphere",50,pos)
image(voi[,,20])

}
\author{
Alessio Veneziano
}
