% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTiffStack.R
\name{writeTiffStack}
\alias{writeTiffStack}
\title{Write TIFF image stack}
\usage{
writeTiffStack(Stack, path, leadname = "stack_", bits = 8, cores = 2)
}
\arguments{
\item{Stack}{an image stack object}

\item{path}{path where the TIFF images will be saved}

\item{cores}{number of cores for parallel computation (default=2)}
}
\description{
Writes a 3D image stack array to a TIFF image stack. It uses the function \code{writeTIFF} in the package 'tiff' and runs in parallel using the packages 'foreach' and 'doParallel'.
}
\author{
Alessio Veneziano
}
