% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBinarized.R
\name{isBinarized}
\alias{isBinarized}
\title{Check binarized image stack}
\usage{
isBinarized(Stack)
}
\arguments{
\item{Stack}{an image stack object}
}
\value{
\code{TRUE} or \code{FALSE} depending on whether \code{Stack} is binarized or not.
}
\description{
Checks if an image stack is binarized (contains only two types of pixels, black and white).
}
\author{
Alessio Veneziano
}
