% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skelLength.R
\name{skelLength}
\alias{skelLength}
\title{Skeleton Length}
\usage{
skelLength(skel)
}
\arguments{
\item{skel}{topological skeleton, output of \code{readAmiraSkeleton}}
}
\value{
\item{edgeL}{a vector of branch lengths, of length equal to the number of branches in the topological skeleton.}
\item{nodeL}{a vector of average branch length per node, of length equal to the number of nodes in the topological skeleton.}
}
\description{
Calculates the branch length of the topological skeleton and the average branch length per node.
}
\examples{
#Calculate branch length of a topological skeleton
data(exampleSkeleton)
len<-skelLength(exampleSkeleton)

}
\references{
Veneziano A, Cazenave M, Alfieri F, Panetta D, Marchi D. 2021. Novel strategies for the characterization of cancellous bone morphology: Virtual isolation and analysis. American Journal of Physical Anthropology.
}
\author{
Alessio Veneziano
}
